+++ node/src/types/kernel/api/block/appendBlock/index.d.ts
@@ -19,3 +19,5 @@
 /* directories */
 
 /* files */
+export * from "./payload";
+export * from "./response";