+++ node/src/types/kernel/api/block/appendBlock/response.d.ts
@@ -16,7 +16,7 @@
  */
 
 /**
- * Prepend blocks
+ * Append blocks
  */
 export interface IResponse {
     /**